refactor static methods
